+++ src/components/ReportList/FilterRegion/SelectRegion/EupmyeondongTable.tsx
@@ -1,22 +1,22 @@
 import React from 'react';
 import { Pressable, StyleSheet, Text, View, FlatList } from 'react-native';
-import COLOR from '../../constants/colors';
+import COLOR from '../../../../constants/colors';
 import FeatherIcon from 'react-native-vector-icons/Feather';
-import { DisasterType } from './types';
+import { SigunguAndEupmyeondongType } from './types';
 
-type SubCategoryTableProps = {
-  selectedDisaster: DisasterType[];
-  setSelectedDisaster: React.Dispatch<React.SetStateAction<DisasterType[]>>;
-  disasterList: DisasterType[];
+type EupmyeondongTableProps = {
+  selectedEupmyeondong: SigunguAndEupmyeondongType[];
+  setSelectedEupmyeondong: React.Dispatch<React.SetStateAction<SigunguAndEupmyeondongType[]>>;
+  eupmyeondongList: SigunguAndEupmyeondongType[];
 };
 
-export default function SubCategory({
-  selectedDisaster,
-  setSelectedDisaster,
-  disasterList,
-}: SubCategoryTableProps) {
-  const handleDisasterItemClick = ({ item }: { item: DisasterType }) => {
-    setSelectedDisaster((prev) => {
+export default function EupmyeondongTable({
+  selectedEupmyeondong,
+  setSelectedEupmyeondong,
+  eupmyeondongList,
+}: EupmyeondongTableProps) {
+  const handleEupmyeondongItemClick = ({ item }: { item: SigunguAndEupmyeondongType }) => {
+    setSelectedEupmyeondong((prev) => {
       if (prev.some((selectedItem) => selectedItem.id === item.id)) {
         return prev.filter((prevItem) => prevItem.id !== item.id);
       } else {
@@ -25,20 +25,20 @@ export default function SubCategory({
     });
   };
 
-  const renderEupmyeondong = ({ item }: { item: DisasterType }) => {
+  const renderEupmyeondong = ({ item }: { item: SigunguAndEupmyeondongType }) => {
     return (
-      <Pressable style={styles.regionItem} onPress={() => handleDisasterItemClick({ item })}>
+      <Pressable style={styles.regionItem} onPress={() => handleEupmyeondongItemClick({ item })}>
         <View style={styles.eupmyeondongItem}>
           <Text
             style={
-              selectedDisaster.some((selectedItem) => selectedItem.id === item.id)
+              selectedEupmyeondong.some((selectedItem) => selectedItem.id === item.id)
                 ? StyleSheet.compose(styles.regionItemtext, styles.selectedEupmyeondongText)
                 : styles.regionItemtext
             }
           >
-            {item.text}
+            {item.singleName}
           </Text>
-          {selectedDisaster.some((selectedItem) => selectedItem.id === item.id) && (
+          {selectedEupmyeondong.some((selectedItem) => selectedItem.id === item.id) && (
             <FeatherIcon name="check" size={18} style={styles.check} />
           )}
         </View>
@@ -48,10 +48,10 @@ export default function SubCategory({
   return (
     <View style={styles.tableRight}>
       <View style={styles.tableTitle}>
-        <Text>소분류</Text>
+        <Text>동,읍,면</Text>
       </View>
       <FlatList
-        data={disasterList}
+        data={eupmyeondongList}
         renderItem={renderEupmyeondong}
         numColumns={1}
         contentContainerStyle={styles.regionList}
