+++ src/components/ReportList/index.js
@@ -1,3 +1,3 @@
 export { default as FilterButtons } from './FilterButtons/FilterButtons';
-export { default as FilterDisasterBottomSheet } from './FilterDisaster/FilterDisasterBottomSheet';
-export { default as FilterRegionBottomSheet } from './FilterRegion/FilterRegionBottomSheet';
+export { default as FilterDisasterBottomSheet } from './FilterDisaster/FilterDisasterBottomSheet/FilterDisasterBottomSheet';
+export { default as FilterRegionBottomSheet } from './FilterRegion/FilterRegionBottomSheet/FilterRegionBottomSheet';
