+++ src/screens/ReportList.tsx
@@ -4,8 +4,8 @@ import { Pressable, StyleSheet, View, ScrollView } from 'react-native';
 import { BottomSheetModal } from '@gorhom/bottom-sheet';
 import COLOR from '../constants/colors';
 import Ionicons from 'react-native-vector-icons/Ionicons';
-import { SigunguAndEupmyeondongType } from '../components/SelectRegion/types';
-import { DisasterType } from '../components/SelectDisaster/types';
+import { SigunguAndEupmyeondongType } from '../components/ReportList/FilterRegion/SelectRegion/types';
+import { DisasterType } from '../components/ReportList/FilterDisaster/SelectDisaster/types';
 import ReportArticleList from '../components/common/ReportArticle/ReportArticleList/ReportArticleList';
 import { AddressBottomSheet } from '../components/common/Modal/BottomSheetModal';
 import TabBar from '../components/common/TabBar/TabBar';
